Refactor for main.rb to only start code and ask for the initial choice
